+++ lib/azure/storage/client_options_error.rb
@@ -22,6 +22,8 @@
 # THE SOFTWARE.
 #--------------------------------------------------------------------------
 
+require 'azure/storage/core'
+
 module Azure::Storage
 
   class InvalidConnectionStringError < Core::StorageError